+++ js/loader.js
@@ -1,6 +1,6 @@
 /*global loadFiles dashtype */
 
-var _DASHTICZ_VERSION=29;
+var _DASHTICZ_VERSION=30;
 var head = document.getElementsByTagName('head')[0],
     script = document.createElement('script');
 script.src = 'dist/bundle.js?t='+_DASHTICZ_VERSION;